Single-Road-Simulation
Initially, the system was modelled in a single-road configuration using the Wokwi Simulator. In this setup, a road is equipped with ultrasonic sensors and LEDs representing traffic lights (Red, Yellow, and Green), as well as a timer that counts down. The traffic density in the system is detected by an ultrasonic sensor, which then adjusts the timing of the traffic light signals accordingly. This simulation validates the basic functions of traffic density detection, appropriate signalling, and dynamic timer control.